Move image_label.py into RimSort.models
